Вводит правки
